Discard in buffer after reading
This is to prevent messages from collecting on the input buffer. This helps during debugging if you are setting breakpoints
and stopping for more than 50ms at a time. Since to poll rate is so slow, this will not affect regular operation of the acceptor.

Also show how to capture the denomination during the OnEscrow event in the test app.