Tested against version 0.7 of OpenNETCF
This uses the serial port alternate and works fine on my system. YMMV. Be sure to grab the dll from http://sourceforge.net/projects/serialportnet/ and add it as a reference to the ApexValidator project. Remove
the using System.IO.SerialPort reference. Make sure that the OpenNETCF dll is in the same directory as your application that you are testing.